Move some skin class into legacy namespace.
