+++ osu.Game.Rulesets.Karaoke/Skinning/Legacy/LegacyColumnBackground.cs
@@ -12,7 +12,7 @@ using osu.Game.Skinning;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Karaoke.Skinning
+namespace osu.Game.Rulesets.Karaoke.Skinning.Legacy
 {
     public class LegacyColumnBackground : LegacyKaraokeColumnElement
     {
