+++ osu.Game.Rulesets.Karaoke/Skinning/Legacy/LegacyKaraokeSkinConfigurationLookup.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Skinning
+namespace osu.Game.Rulesets.Karaoke.Skinning.Legacy
 {
     public class LegacyKaraokeSkinConfigurationLookup
     {
