+++ osu.Game.Rulesets.Karaoke/Skinning/Legacy/LegacyStageBackground.cs
@@ -10,7 +10,7 @@ using osu.Game.Rulesets.UI.Scrolling;
 using osu.Game.Skinning;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Skinning
+namespace osu.Game.Rulesets.Karaoke.Skinning.Legacy
 {
     public class LegacyStageBackground : LegacyKaraokeElement
     {
